requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==4.0.0
argcomplete==3.1.2
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blis==0.7.10
boto3==1.28.69
botocore==1.31.69
Brotli==1.1.0
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
confection==0.1.3
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
dacite==1.7.0
dataclasses-json==0.5.14
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
dnspython==2.4.2
executing==1.2.0
fiftyone==0.22.2
fiftyone-brain==0.13.2
fiftyone-db==0.4.0
filelock==3.13.1
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.1
future==0.18.3
glob2==0.7
graphql-core==3.2.3
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.17.3
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
imageio==2.31.6
inflate64==0.3.1
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonlines==4.0.0
kaleido==0.2.1
kiwisolver==1.4.5
langchain==0.0.281
langcodes==3.3.0
langsmith==0.0.33
lazy_loader==0.3
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mongoengine==0.24.2
motor==3.3.1
mpmath==1.3.0
multidict==6.0.4
multivolumefile==0.2.3
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.2
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
openai==0.28.0
opencv-python==4.8.0.76
opencv-python-headless==4.8.1.78
packaging==23.1
pandas==2.1.1
parso==0.8.3
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
plotly==5.17.0
pprintpp==0.4.0
preshed==3.0.9
priority==2.0.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py7zr==0.20.6
pybcj==1.0.1
pycryptodomex==3.19.0
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.16.1
pymongo==4.5.0
pyparsing==3.1.1
pyppmd==1.0.0
python-dateutil==2.8.2
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzstd==0.15.9
rarfile==4.1
regex==2023.10.3
requests==2.31.0
retrying==1.3.4
s3transfer==0.7.0
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.20
srsly==2.4.7
sse-starlette==0.10.3
sseclient-py==1.8.0
stack-data==0.6.2
starlette==0.31.1
strawberry-graphql==0.138.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
texttable==1.7.0
thinc==8.1.12
threadpoolctl==3.2.0
tifffile==2023.9.26
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
traitlets==5.9.0
transformers==4.35.0
triton==2.1.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
universal-analytics-python3==1.1.1
urllib3==2.0.4
voxel51-eta==0.12.0
wasabi==1.1.2
wcwidth==0.2.6
wrapt==1.15.0
wsproto==1.2.0
xmltodict==0.13.0
yarl==1.9.2
